+++ src/components/DevtoolsPane/MicDeviceForm.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import type React from 'react'
 import { FormGroup } from 'react-bootstrap'
 
 import { setMicDevice } from '@/app/actions'
@@ -8,9 +8,10 @@ import { TooltipFormCheck } from './TooltipFormCheck'
 
 export const MicDeviceForm: React.FC = () => {
   const micDevice = useAppSelector((state) => state.micDevice)
+  const connectionStatus = useAppSelector((state) => state.soraContents.connectionStatus)
   const sora = useAppSelector((state) => state.soraContents.sora)
   const audio = useAppSelector((state) => state.audio)
-  const disabled = !(sora ? sora.audio : audio)
+  const disabled = !(sora && connectionStatus === 'connected' ? sora.audio : audio)
   const dispatch = useAppDispatch()
   const onChange = (event: React.ChangeEvent<HTMLInputElement>): void => {
     dispatch(setMicDevice(event.target.checked))