+++ src/components/DevtoolsPane/RequestMediaButton.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import type React from 'react'
 
 import { requestMedia } from '@/app/actions'
 import { useAppDispatch, useAppSelector } from '@/app/hooks'