+++ src/components/DevtoolsPane/SignalingUrlCandidatesForm.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import type React from 'react'
 import { Col, FormControl, FormGroup, Row } from 'react-bootstrap'
 
 import { setEnabledSignalingUrlCandidates, setSignalingUrlCandidates } from '@/app/actions'