+++ src/components/DevtoolsPane/VideoVP9ParamsForm.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import type React from 'react'
 import { Col, FormControl, FormGroup, Row } from 'react-bootstrap'
 
 import { setEnabledVideoVP9Params, setVideoVP9Params } from '@/app/actions'