+++ src/types.ts
@@ -10,7 +10,7 @@ import type {
   TransportType,
 } from 'sora-js-sdk'
 
-import {
+import type {
   ASPECT_RATIO_TYPES,
   AUDIO_BIT_RATES,
   AUDIO_CODEC_TYPES,
@@ -41,6 +41,12 @@ import {
   VIDEO_CODEC_TYPES,
   VIDEO_CONTENT_HINTS,
 } from '@/constants'
+
+export type RemoteClient = {
+  mediaStream: MediaStream
+  clientId: string | null
+  connectionId: string
+}
 
 export type SoraDevtoolsState = {
   alertMessages: AlertMessage[]
@@ -99,7 +105,7 @@ export type SoraDevtoolsState = {
     clientId: string | null
     sessionId: string | null
     localMediaStream: MediaStream | null
-    remoteMediaStreams: MediaStream[]
+    remoteClients: RemoteClient[]
     prevStatsReport: RTCStats[]
     statsReport: RTCStats[]
     datachannels: DataChannelConfiguration[]
@@ -197,7 +203,7 @@ export type Json =
 
 // HTMLVideoElement interface に setSinkId を追加
 export interface CustomHTMLVideoElement extends HTMLVideoElement {
-  setSinkId(audioId: string): void
+  setSinkId(audioId: string): Promise<void>
 }
 
 // RTCMediaStreamTrackStats に jitterBuffer 関連を追加